UserDetails dump/restore
